workshop/01-dlt/pyproject.toml
[project]
name = "01-dlt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=6.0.0",
    "dlt[workspace]>=1.22.2",
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.20.2",
]
